requirements.txt
Pillow==8.1.2
streamlit==1.29.0
